Extract AlbescentSelectCard to its own file, byte-for-byte (#2329)
The eighth and last archetype leaves the 595-line `FactionSelectCard.tsx`
for `components/selectCard/AlbescentSelectCard.tsx`, matching how
`components/taskCard/` is laid out. The dispatcher now holds only the
shared types, `LEGACY_SLUG` and `FactionSelectCard` itself. The manifest
row is unchanged in shape — `lazyArchetype(() => import(...))` points at
the new module and drops the `.then((m) => ({ default: m.Albescent... }))`
unwrap, because the new file default-exports it the way every task card
does.

THE REPAINT IS PARKED, NOT FORGOTTEN. Unlike its seven siblings this tile
gathers no register: `AlbescentTaskCard` is fifty lines that draw nothing
of their own — `DefaultTaskCard` plus two motion overlays, per ADR-0048 —
so there is no register to gather, and its families are
`--faction-default-*`. Whether the tile keeps painting in
`--albescent-reveal-*` is contested on #2301, which has ruled the reveal
family gets a dark half while index.css still forbids exactly that. Not a
colour moves here.

PROOF THE DIRECTORY IS UNCHANGED. `renderToStaticMarkup` across all six
state x members combinations, before and after the move: identical, same
md5. And through the dispatcher, `<FactionSelectCard faction="albescent">`
resolves to the new module and is byte-identical to the component rendered
directly, on all six.
